Consolidate AzureAppService into Azure
