+++ source/Calamari.Azure/AppServices/Azure/AccountVariables.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Calamari.AzureAppService.Azure
+namespace Calamari.Azure.AppServices.Azure
 {
     public static class AccountVariables
     {
