+++ source/Calamari.Azure/AppServices/Azure/SpecialVariables.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Calamari.AzureAppService.Azure
+namespace Calamari.Azure.AppServices.Azure
 {
     static class SpecialVariables
     {
