+++ source/Calamari.Azure/AppServices/Behaviors/AzureAppServiceAppSettingsBehaviour.cs
@@ -6,18 +6,17 @@ using System.Linq;
 using System.Threading.Tasks;
 using Azure.ResourceManager;
 using Azure.ResourceManager.AppService.Models;
-using Calamari.Azure;
-using Calamari.Azure.AppServices;
-using Calamari.AzureAppService.Azure;
+using Calamari.Azure.AppServices.Azure;
+using Calamari.AzureAppService;
 using Calamari.CloudAccounts;
 using Calamari.Common.Commands;
 using Calamari.Common.Plumbing.Logging;
 using Calamari.Common.Plumbing.Pipeline;
 using Newtonsoft.Json;
 using Octopus.CoreUtilities.Extensions;
-using AccountVariables = Calamari.AzureAppService.Azure.AccountVariables;
+using AccountVariables = Calamari.Azure.AppServices.Azure.AccountVariables;
 
-namespace Calamari.AzureAppService.Behaviors
+namespace Calamari.Azure.AppServices.Behaviors
 {
     class AzureAppServiceSettingsBehaviour : IDeployBehaviour
     {
