+++ source/Calamari.Azure/AppServices/PublishingProfile.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Xml.Linq;
 
-namespace Calamari.AzureAppService
+namespace Calamari.Azure.AppServices
 {
     class PublishingProfile
     {
