+++ source/Calamari.Azure/AppServices/TargetDiscoveryCommand.cs
@@ -1,9 +1,10 @@
-﻿using System.Collections.Generic;
-using Calamari.AzureAppService.Behaviors;
+﻿using System;
+using System.Collections.Generic;
+using Calamari.Azure.AppServices.Behaviors;
 using Calamari.Common.Commands;
 using Calamari.Common.Plumbing.Pipeline;
 
-namespace Calamari.AzureAppService
+namespace Calamari.Azure.AppServices
 {
     [Command("target-discovery", Description = "Discover Azure web applications")]
     public class TargetDiscoveryCommand : PipelineCommand
