+++ source/Calamari.Azure/Program.cs
@@ -1,19 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Net;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using Calamari.Common;
 using Calamari.Common.Plumbing.Logging;
 
-
-namespace Calamari.AzureAppService
+namespace Calamari.Azure
 {
     public class Program : CalamariFlavourProgramAsync
     {
-        public Program(ILog log) : base(log)
-        {
-        }
+        public Program(ILog log)
+            : base(log)
+        { }
 
         public static Task<int> Main(string[] args)
         {
