+++ source/Calamari.AzureAppService.Tests/AppServiceBehaviourFixture.cs
@@ -15,7 +15,9 @@ using Azure.ResourceManager.AppService.Models;
 using Azure.ResourceManager.Resources;
 using Azure.ResourceManager.Storage;
 using Azure.ResourceManager.Storage.Models;
-using Calamari.AzureAppService.Azure;
+using Calamari.Azure;
+using Calamari.Azure.AppServices;
+using Calamari.Azure.AppServices.Azure;
 using Calamari.Common.FeatureToggles;
 using Calamari.Common.Plumbing.FileSystem;
 using Calamari.Common.Plumbing.Variables;
