+++ source/Calamari.AzureAppService.Tests/AppServiceSettingsBehaviourFixture.cs
@@ -6,9 +6,8 @@ using Azure;
 using Azure.ResourceManager.AppService.Models;
 using Azure.ResourceManager.Resources;
 using Calamari.Azure.AppServices;
-using Calamari.AzureAppService.Azure;
-using Calamari.AzureAppService.Behaviors;
-using Calamari.AzureAppService.Json;
+using Calamari.Azure.AppServices.Azure;
+using Calamari.Azure.AppServices.Behaviors;
 using Calamari.Common.Commands;
 using Calamari.Common.Plumbing.Variables;
 using Calamari.Testing.Helpers;
