+++ source/Calamari.AzureAppService.Tests/AzureAppServiceDeployContainerBehaviourFixture.cs
@@ -8,8 +8,8 @@ using Azure.ResourceManager.AppService;
 using Azure.ResourceManager.AppService.Models;
 using Azure.ResourceManager.Resources;
 using Calamari.Azure.AppServices;
-using Calamari.AzureAppService.Azure;
-using Calamari.AzureAppService.Behaviors;
+using Calamari.Azure.AppServices.Azure;
+using Calamari.Azure.AppServices.Behaviors;
 using Calamari.Common.Commands;
 using Calamari.Common.Plumbing.Variables;
 using Calamari.Testing.Helpers;
