+++ source/Calamari.AzureAppService.Tests/Calamari.AzureAppService.Tests.csproj
@@ -22,7 +22,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Calamari.AzureAppService\Calamari.AzureAppService.csproj" />
+    <ProjectReference Include="..\Calamari.Azure\Calamari.Azure.csproj" />
     <ProjectReference Include="..\Calamari.Testing\Calamari.Testing.csproj" />
   </ItemGroup>
