+++ source/Calamari.sln
@@ -42,8 +42,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "_build", "..\build\_build.c
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.Azure", "Calamari.Azure\Calamari.Azure.csproj", "{AC3A676B-D720-4AC8-A188-B25DB72969CB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.AzureAppService", "Calamari.AzureAppService\Calamari.AzureAppService.csproj", "{0A1E26DB-E41C-4186-882A-75BA3EC10490}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.AzureAppService.Tests", "Calamari.AzureAppService.Tests\Calamari.AzureAppService.Tests.csproj", "{11AFAB0B-A3EE-4FC1-95A7-FB480B3C938B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calamari.ConsolidateCalamariPackages.Tests", "Calamari.ConsolidateCalamariPackages.Tests\Calamari.ConsolidateCalamariPackages.Tests.csproj", "{02FF12AA-8807-4041-A59A-04CBE91AA1E1}"
@@ -124,10 +122,6 @@ Global
 		{AC3A676B-D720-4AC8-A188-B25DB72969CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AC3A676B-D720-4AC8-A188-B25DB72969CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AC3A676B-D720-4AC8-A188-B25DB72969CB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0A1E26DB-E41C-4186-882A-75BA3EC10490}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0A1E26DB-E41C-4186-882A-75BA3EC10490}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0A1E26DB-E41C-4186-882A-75BA3EC10490}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0A1E26DB-E41C-4186-882A-75BA3EC10490}.Release|Any CPU.Build.0 = Release|Any CPU
 		{11AFAB0B-A3EE-4FC1-95A7-FB480B3C938B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{11AFAB0B-A3EE-4FC1-95A7-FB480B3C938B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{11AFAB0B-A3EE-4FC1-95A7-FB480B3C938B}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -208,14 +202,14 @@ Global
 		{D8DEC40C-948F-4806-AE87-1A7502E41A06}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D8DEC40C-948F-4806-AE87-1A7502E41A06}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D8DEC40C-948F-4806-AE87-1A7502E41A06}.Release|Any CPU.Build.0 = Release|Any CPU
-        {74F2739F-A501-4F30-BF9E-75A087C3B917}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-        {74F2739F-A501-4F30-BF9E-75A087C3B917}.Debug|Any CPU.Build.0 = Debug|Any CPU
-        {74F2739F-A501-4F30-BF9E-75A087C3B917}.Release|Any CPU.ActiveCfg = Release|Any CPU
-        {74F2739F-A501-4F30-BF9E-75A087C3B917}.Release|Any CPU.Build.0 = Release|Any CPU
-        {A3BFFA0E-7514-41DE-A141-E124133D0665}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-        {A3BFFA0E-7514-41DE-A141-E124133D0665}.Debug|Any CPU.Build.0 = Debug|Any CPU
-        {A3BFFA0E-7514-41DE-A141-E124133D0665}.Release|Any CPU.ActiveCfg = Release|Any CPU
-        {A3BFFA0E-7514-41DE-A141-E124133D0665}.Release|Any CPU.Build.0 = Release|Any CPU
+								{74F2739F-A501-4F30-BF9E-75A087C3B917}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+								{74F2739F-A501-4F30-BF9E-75A087C3B917}.Debug|Any CPU.Build.0 = Debug|Any CPU
+								{74F2739F-A501-4F30-BF9E-75A087C3B917}.Release|Any CPU.ActiveCfg = Release|Any CPU
+								{74F2739F-A501-4F30-BF9E-75A087C3B917}.Release|Any CPU.Build.0 = Release|Any CPU
+								{A3BFFA0E-7514-41DE-A141-E124133D0665}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+								{A3BFFA0E-7514-41DE-A141-E124133D0665}.Debug|Any CPU.Build.0 = Debug|Any CPU
+								{A3BFFA0E-7514-41DE-A141-E124133D0665}.Release|Any CPU.ActiveCfg = Release|Any CPU
+								{A3BFFA0E-7514-41DE-A141-E124133D0665}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B4EB8110-50D8-4C7E-8DBF-102AE57FD943}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B4EB8110-50D8-4C7E-8DBF-102AE57FD943}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B4EB8110-50D8-4C7E-8DBF-102AE57FD943}.Release|Any CPU.ActiveCfg = Release|Any CPU
